+++ lib/animation/animation_player.dart
@@ -1,0 +1,3 @@
+class AnimationPlayer {
+
+}